+++ src/services/discord/core.service.ts
@@ -9,6 +9,8 @@ import channelService from './channel.service';
 import roleService from './role.service';
 import guildMemberService from './guildMember.service';
 import { discord } from '../../config/oAtuh2';
+import httpStatus from 'http-status';
+
 const logger = parentLogger.child({ module: 'DiscordCoreService' });
 
 /**
@@ -129,16 +131,15 @@ async function leaveBotFromGuild(guildId: Snowflake) {
 /**
  * exchange discord code with access token
  * @param {string} code
-   @param {string} redirect_uri
  * @returns {Promise<IDiscordOAuth2EchangeCode>}
  */
-async function exchangeCode(code: string, redirect_uri: string): Promise<IDiscordOAuth2EchangeCode> {
+async function exchangeCode(code: string): Promise<IDiscordOAuth2EchangeCode> {
   try {
     const data = {
       client_id: config.oAuth2.discord.clientId,
       client_secret: config.oAuth2.discord.clientSecret,
       grant_type: 'authorization_code',
-      redirect_uri,
+      redirect_uri: config.oAuth2.discord.callbackURI.connect,
       code,
     };
 
@@ -155,7 +156,7 @@ async function exchangeCode(code: string, redirect_uri: string): Promise<IDiscor
       throw new Error(`Failed to exchange discord code: ${errorResponse}`);
     }
   } catch (error) {
-    logger.error({ redirect_uri, error }, 'Failed to exchange discord code');
+    logger.error({ redirect_uri: config.oAuth2.discord.callbackURI.connect, error }, 'Failed to exchange discord code');
     throw new ApiError(590, 'Can not fetch from discord API');
   }
 }
@@ -312,6 +313,36 @@ class DiscordBotManager {
   }
 }
 
+/**
+ * refresh token
+ * @param {string} refreshToken
+ * @returns {Promise<IDiscordOAuth2EchangeCode>}
+ */
+async function refreshAuth(refreshToken: string): Promise<IDiscordOAuth2EchangeCode> {
+  try {
+    const data = {
+      client_id: config.oAuth2.discord.clientId,
+      client_secret: config.oAuth2.discord.clientSecret,
+      grant_type: 'refresh_token',
+      refresh_token: refreshToken,
+    };
+
+    const response = await fetch('https://discord.com/api/oauth2/token', {
+      method: 'POST',
+      body: new URLSearchParams(data),
+      headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
+    });
+    if (response.ok) {
+      return await response.json();
+    } else {
+      throw new Error();
+    }
+  } catch (error) {
+    logger.error({ error }, 'Failed to refresh discord auth');
+    throw new ApiError(httpStatus.INTERNAL_SERVER_ERROR, 'Can not fetch from discord API');
+  }
+}
+
 export default {
   exchangeCode,
   getUserFromDiscordAPI,
@@ -323,4 +354,5 @@ export default {
   getRequirePermissionsForModule,
   getCombinedPermissionsValue,
   getPermissionsStatus,
+  refreshAuth
 };
